+++ src/agent/agent.ts
@@ -6,7 +6,7 @@ import { initializeDatabase } from '@/common/db'
 import { isNull } from '@/common/functions'
 import { PathResolver } from '@/common/path-resolver'
 import { AgentcoinRuntime } from '@/common/runtime'
-import { AyaOSOptions, Context, ContextHandler, ModelConfig, SdkEventKind } from '@/common/types'
+import { Context, ContextHandler, ModelConfig, SdkEventKind } from '@/common/types'
 import agentcoinPlugin from '@/plugins/agentcoin'
 import { AgentcoinService } from '@/services/agentcoinfun'
 import { ConfigService } from '@/services/config'
@@ -49,16 +49,22 @@ export class Agent implements IAyaAgent {
   private runtime_: AgentcoinRuntime | undefined
   private pathResolver: PathResolver
   private keychainService: KeychainService
-  private matchThreshold?: number
+  public matchThreshold?: number
+  public matchLimit?: number
 
-  constructor(options?: AyaOSOptions) {
+  constructor(options?: {
+    modelConfig?: ModelConfig
+    dataDir?: string
+    knowledge?: { matchThreshold?: number; matchLimit?: number }
+  }) {
     this.modelConfig = options?.modelConfig
     if (reservedAgentDirs.has(options?.dataDir)) {
       throw new Error('Data directory already used. Please provide a unique data directory.')
     }
     reservedAgentDirs.add(options?.dataDir)
     this.pathResolver = new PathResolver(options?.dataDir)
-    this.matchThreshold = options?.matchThreshold
+    this.matchThreshold = options?.knowledge?.matchThreshold
+    this.matchLimit = options?.knowledge?.matchLimit
   }
 
   get runtime(): AgentcoinRuntime {
@@ -85,6 +91,11 @@ export class Agent implements IAyaAgent {
   }
 
   async start(): Promise<void> {
+    if (isNull(process.env.POSTGRES_URL)) {
+      elizaLogger.error('POSTGRES_URL is not set, please set it in your .env file')
+      process.exit(1)
+    }
+
     let runtime: AgentcoinRuntime | undefined
 
     try {
@@ -99,11 +110,6 @@ export class Agent implements IAyaAgent {
         this.pathResolver
       )
       await agentcoinService.provisionIfNeeded()
-
-      if (isNull(process.env.POSTGRES_URL)) {
-        elizaLogger.error('POSTGRES_URL is not set, please set it in your .env file')
-        process.exit(1)
-      }
 
       // eagerly start event service
       const agentcoinCookie = await agentcoinService.getCookie()
@@ -161,7 +167,8 @@ export class Agent implements IAyaAgent {
           agentId: character.id
         },
         pathResolver: this.pathResolver,
-        matchThreshold: this.matchThreshold
+        matchThreshold: this.matchThreshold,
+        matchLimit: this.matchLimit
       })
       this.runtime_ = runtime
 
@@ -228,12 +235,13 @@ export class Agent implements IAyaAgent {
       })
 
       this.runtime.clients = await initializeClients(this.runtime.character, this.runtime)
-      this.register('service', knowledgeService)
-      this.register('service', memoriesService)
-      this.register('service', walletService)
+      await Promise.all([
+        this.register('service', knowledgeService),
+        this.register('service', memoriesService),
+        this.register('service', walletService)
+      ])
       // no need to await these. it'll lock up the main process
-      void knowledgeService.start()
-      void configService.start()
+      void Promise.all([configService.start(), knowledgeService.start()])
 
       elizaLogger.info(`Started ${this.runtime.character.name} as ${this.runtime.agentId}`)
     } catch (error: unknown) {
@@ -259,18 +267,18 @@ export class Agent implements IAyaAgent {
     )
   }
 
-  register(kind: 'service', handler: Service): void
-  register(kind: 'provider', handler: Provider): void
-  register(kind: 'action', handler: Action): void
-  register(kind: 'plugin', handler: Plugin): void
-  register(kind: 'evaluator', handler: Evaluator): void
+  async register(kind: 'service', handler: Service): Promise<void>
+  async register(kind: 'provider', handler: Provider): Promise<void>
+  async register(kind: 'action', handler: Action): Promise<void>
+  async register(kind: 'plugin', handler: Plugin): Promise<void>
+  async register(kind: 'evaluator', handler: Evaluator): Promise<void>
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  register(kind: string, handler: any): void {
+  async register(kind: string, handler: any): Promise<void> {
     switch (kind) {
       case 'service':
         this.services.push(handler)
         if (this.runtime_) {
-          void this.runtime.registerService(handler)
+          await this.runtime.registerService(handler)
         }
         break
       case 'action':